requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.7
psycopg[binary]
gunicorn==21.2.0
